% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scores}
\alias{scores}
\title{Calculation of Probabilities (Not For End User)}
\usage{
scores(initsa=NULL,theta=NULL)
}
\arguments{
\item{initsa}{Model SA.}

\item{theta}{Optional(Conditional Prababilities).}
}
\description{
This function is not for end user.
}
\examples{
\dontrun{
scores(initsa)
}




}
