% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOC.R
\name{TOC.sampleData}
\alias{TOC.sampleData}
\title{Learning (Not For End User)}
\usage{
TOC.sampleData(initsa=NULL,n)
}
\arguments{
\item{initsa}{SA Model.}

\item{n}{Length of input sample set sequence.}
}
\description{
This function is not for end user.
}
\examples{
\dontrun{
states<-c('s1','s2')
inputSymbols<-c('a','b')
outputSymbols<-c(0,1)
transProb<-matrix(c(0.70,0.50, 0.30,0.50), nrow = 2, ncol = 2,byrow = TRUE)
emissionProb<-matrix(c(0.50,0.30, 0.40,0.60,.50,.70,.60,.40), nrow = 2, ncol = 4, byrow = TRUE)
initsa<-initsa(states,inputSymbols,outputSymbols,emissionProb,transProb)
n<-3
TOC.sampleData(initsa, n)



}
}
